+++ extensions/Worker.Extensions.EventHubs/src/Constants.cs
@@ -1,7 +1,7 @@
 // Copyright (c) .NET Foundation. All rights reserved.
 // Licensed under the MIT License. See License.txt in the project root for license information.
 
-namespace Microsoft.Azure.Functions.Worker.Extensions.ServiceBus
+namespace Microsoft.Azure.Functions.Worker.Extensions.EventHubs
 {
     internal static class Constants
     {